add test to navigator
