+++ VideoWeb/VideoWeb/ClientApp/src/app/home/navigator/navigator.component.html
@@ -1,1 +1,1 @@
-<p>navigator works!</p>
+<p class="govuk-body govuk-!-padding-top-2">Please wait whilst we log you in...</p>
