+++ VideoWeb/VideoWeb/ClientApp/src/app/home/navigator/navigator.component.ts
@@ -9,7 +9,6 @@ import { UserProfileResponse, Role } from '../../services/clients/api-client';
 @Component({
   selector: 'app-navigator',
   templateUrl: './navigator.component.html',
-  styleUrls: ['./navigator.component.css']
 })
   export class NavigatorComponent implements OnInit {
     constructor(
